backend/pyproject.toml
[tool.poetry]
name = "swift-demo"
version = "0.0.1"
description = ""
authors = ["Kyle Getrost <kyle@motionhungry.com>"]

[tool.poetry.dependencies]
python = "^3.10"
load-dotenv = "^0.1.0"
requests = "^2.28.2"
PyJWT = "^2.6.0"
cryptography = "^39.0.0"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
